requirements.txt
requests
pandas
httpx
bs4
feedparser
python-dotenv
cot-reports
pytz
